refactor: changed files and class names
