+++ app/services/metrics.py
@@ -5,8 +5,8 @@ from app.services.utils.collection import (
     get_items_asset_url,
     get_asset_href_by_item_id,
 )
-from app.routes.schemas.polygon import PolygonGeometry
-from app.routes.schemas.MetricValues import MetricResponse
+from app.routes.schemas.PolygonRequest import PolygonGeometry
+from app.routes.schemas.MetricResponse import MetricResponse
 from app.services.utils.metadata import fetch_collection_metadata
 from app.services.utils.metrics_config import metric_group_key
